changed position to floats

+++ src/PlayerTomcat.java
@@ -1,7 +1,7 @@
 public class PlayerTomcat extends Entity{
 	//z should always be 0 for the player
 	
-	public PlayerTomcat(int nX, int nY, int nZ, int nxV, int nyV, int nzV) {
+	public PlayerTomcat(float nX, float nY, float nZ, float nxV, float nyV, float nzV) {
 		super(nX, nY, nZ);
 	}
 }
